# auth_backend/interfaces/openapi.runtime.json
{"openapi":"3.1.0","info":{"title":"Secure User Authentication System API","description":"FastAPI backend providing user registration and login with PostgreSQL storage and JWT authentication.","version":"0.1.0"},"paths":{"/":{"get":{"tags":["System"],"summary":"Health Check","description":"Simple health check endpoint used for container readiness/liveness probes.","operationId":"health_check","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/signup":{"post":{"tags":["Authentication"],"summary":"Register a new user","description":"Creates a new user account with a unique email and a bcrypt-hashed password.","operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"400":{"description":"Validation error or weak password.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Email already exists.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Unexpected server error.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/login":{"post":{"tags":["Authentication"],"summary":"Login","description":"Validates credentials (email + password) against stored user records and returns a signed JWT access token.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"401":{"description":"Invalid credentials.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Invalid request.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AuthResponse":{"properties":{"access_token":{"type":"string","title":"Access Token","description":"JWT access token to authenticate subsequent requests."},"token_type":{"type":"string","title":"Token Type","description":"Token type (always 'bearer').","default":"bearer"},"email":{"type":"string","format":"email","title":"Email","description":"User email."},"user_id":{"type":"integer","title":"User Id","description":"User id."},"message":{"type":"string","title":"Message","description":"Human-readable success message for the client UI."}},"type":"object","required":["access_token","email","user_id","message"],"title":"AuthResponse","description":"Response returned after successful signup/login."},"ErrorResponse":{"properties":{"detail":{"type":"string","title":"Detail","description":"Human-readable error message."},"code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Code","description":"Optional stable error code for clients."}},"type":"object","required":["detail"],"title":"ErrorResponse","description":"Standard error response."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email","description":"Registered email address.","examples":["user@example.com"]},"password":{"type":"string","title":"Password","description":"User password.","examples":["Str0ngPassw0rd"]}},"type":"object","required":["email","password"],"title":"LoginRequest","description":"Request body for login."},"SignupRequest":{"properties":{"name":{"type":"string","maxLength":200,"minLength":1,"title":"Name","description":"User full name.","examples":["Jane Doe"]},"phone":{"type":"string","maxLength":40,"minLength":0,"title":"Phone","description":"User phone number.","default":"N/A","examples":["+1 555 123 4567"]},"email":{"type":"string","format":"email","title":"Email","description":"User email address (must be unique).","examples":["user@example.com"]},"password":{"type":"string","minLength":8,"title":"Password","description":"Password (min 8 chars, must contain uppercase, lowercase, and number).","examples":["Str0ngPassw0rd"]}},"type":"object","required":["name","email","password"],"title":"SignupRequest","description":"Request body for user registration.\n\nCanonical request body (documented in OpenAPI):\n    {\n      \"name\": \"Jane Doe\",\n      \"phone\": \"+1 555 123 4567\",\n      \"email\": \"user@example.com\",\n      \"password\": \"Str0ngPassw0rd\"\n    }\n\nCompatibility notes (to prevent preview 422s due to client mismatch):\n- Accepts `fullName` as an alias for `name`.\n- Accepts `phoneNumber` as an alias for `phone`.\n- Accepts a few common alternative keys used by other templates/forms:\n  - `username` as alias for `name`\n  - `pass` as alias for `password`\n- Ignores unknown extra fields (e.g., `confirmPassword`) rather than 422-ing.\n- If `phone` is omitted (or empty), it is defaulted to \"N/A\" (still stored in DB).\n  This preserves the DB's NOT NULL constraint while keeping API tolerant."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}},"tags":[{"name":"System","description":"Health checks and operational endpoints."},{"name":"Authentication","description":"User signup and login endpoints."}]}